Crate app

metw-accounts-center app

This crate contains the application binary, bootstrapping all handlers to the axum routers.

Note: The Axum application is intended to run under a reverse proxy. The reverse proxy must provide the remote address via the X-Real-IP header.

Modules

middleware

API middleware.

res

API results.

routes

API routes.

Structs

ApiDoc

OpenAPI documentation.

Functions

app

Constructs the web API.