Function app::app_sensor::convert_sensor_data

fn convert_sensor_data(
    sensor_data: sensor_data_ptr, 
    sensor_type: sensor_type_t
) -> SensorValue

Convert the raw temperature value received from Mynewt into a SensorValue for transmission, which includes the sensor data key t. sensor_type indicates the type of data in sensor_data.