Struct app::backend::db_ops::DBOps

pub struct DBOps {
    db_conn: Connection,
    db_file: String,
    file_nr: i32,
}

Fields
db_conn: Connectiondb_file: Stringfile_nr: i32
Implementations

impl DBOps

pub fn new(db_file: &str) -> Result<Self, Error>

Returns a new DBOps struct with a connection to the specified database file and initializes the table if it does not exist.

Arguments

db_file - The path to the database file.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert_eq!(db_ops.db_conn, Connection::open("signatures.db").unwrap());

pub fn init_table(&self) -> Result<(), Error>

Initializes the signatures table if it does not exist.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert!(db_ops.init_table().is_ok());

pub fn update_db(&mut self)

Updates the database by downloading any missing files and inserting their hashes into the signatures table.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let mut db_ops = DBOps::new("signatures.db").unwrap();
db_ops.update_db();

pub fn download_files(&mut self, files: Vec<i32>)

Downloads the specified files and inserts their hashes into the signatures table.

Examples

use rusqlite::Connection; 
use virus_scanner::backend::db_ops::DBOps; 
let mut db_ops = DBOps::new("signatures.db").unwrap(); 
db_ops.download_files(vec![1, 2, 3]); 

pub fn download_file(
file_nr: i32
) -> Result<Option<Vec<(String, String)>>, Error>

Downloads the specified file and returns its content and file number as a tuple in the form of (file_nr, content). Returns None if the file does not exist or if there was an error creating a String from the file’s bytes.

Examples

use rusqlite::Connection; 
use virus_scanner::backend::db_ops::DBOps;
let file = DBOps::download_file(1).unwrap(); 
assert!(file.is_some()); 

pub fn insert_hashes(
&mut self,
hashes: Vec<(String, String)>
) -> Result<(), Error>

Inserts the given hashes into the signatures table.

Examples

use rusqlite::Connection; 
use virus_scanner::backend::db_ops::DBOps; 
let mut db_ops = DBOps::new("signatures.db").unwrap(); 
db_ops.insert_hashes(vec![("abcdef".to_owned(), "1".to_owned())]).unwrap(); 

pub fn hash_exists(&self, hash_str: &str) -> Result<bool, Error>

Returns true or false depending on if the given hash gets found in the database

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert_eq!(db_ops.hash_exists("abcd1234").unwrap(), false);

pub fn count_hashes(&self) -> Result<u64, Error>

Returns the number of hashes in the signatures table.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert_eq!(db_ops.count_hashes().unwrap(), 0);

pub fn _remove_hash(&self, hash_str: &str) -> Result<(), Error>

Removes the specified hash from the signatures table.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert!(db_ops._remove_hash("abcd1234").is_ok());

pub fn get_file_list(&self) -> i32

Returns the file numbers of the files that are present online.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert!(db_ops.get_file_list() > 0);

pub fn file_exists(file_nr: i32) -> Result<bool, Error>

Returns whether the file with the specified file number exists online.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert!(db_ops.file_exists(123).unwrap_or(false));

pub fn get_db_files(&self) -> Option<Vec<i32>>

Returns a vector of the file numbers of the files that are present in the signatures table.

Examples

use rusqlite::Connection;
use virus_scanner::backend::db_ops::DBOps;
let db_ops = DBOps::new("signatures.db").unwrap();
assert!(db_ops.get_db_files().is_some());

pub fn get_diff_file(&self) -> Vec<i32>

Returns a list of file numbers for which there are no corresponding hashes in the signatures table.

Examples

use rusqlite::Connection; 
use virus_scanner::backend::db_ops::DBOps; 
let db_ops = DBOps::new("signatures.db").unwrap(); 
assert!(db_ops.get_diff_file().len() >= 0);