Function app::start_scanner

pub(crate) async fn start_scanner(
    path: String,
    update: bool,
    dbfile: Option<String>
) -> Result<(), String>

Starts the scanner for the given path and updates the database if update is true.

Arguments

path - The path to scan.

update - Whether to update the database before scanning.

dbfile - An optional path to a specific database file.

Returns

An empty Result object if the scanner was successfully started, or an Err with an error message if an error occurred.