app::stores::date
Struct DateMapper

pub struct DateMapper {
    years: RwLock<HashSet<String>>,
    months: RwLock<HashMap<String, HashSet<String>>>,
    days: RwLock<HashMap<String, HashSet<String>>>,
    objects: RwLock<HashMap<String, HashSet<String>>>,
}

used to keep lookups

Fields
years: RwLock<HashSet<String>>

the years

months: RwLock<HashMap<String, HashSet<String>>>

the months the key of the hashmap is the ${year}“;

days: RwLock<HashMap<String, HashSet<String>>>

the days the key of the hashmap is the ${year}-{month}“;

objects: RwLock<HashMap<String, HashSet<String>>>

the objects the key of the hashmap is the ${year}-{month}-{day}“;

Implementations

impl DateMapper

pub fn years(&self) -> Result<Vec<String>, ExogenesisError>

get years

pub fn months(&self, year: &str) -> Result<Vec<String>, ExogenesisError>

pub fn days( &self, year: &str, month: &str, ) -> Result<Vec<String>, ExogenesisError>

pub fn objects( &self, year: &str, month: &str, day: &str, ) -> Result<Vec<String>, ExogenesisError>

pub fn add_years(&mut self, years: Vec<String>) -> Result<(), ExogenesisError>

add new years data

pub fn add_months( &mut self, year: &str, months: Vec<String>, ) -> Result<(), ExogenesisError>

add new months data

pub fn add_days( &mut self, year: &str, month: &str, days: Vec<String>, ) -> Result<(), ExogenesisError>

add new months data

pub fn add_objects( &mut self, year: &str, month: &str, day: &str, objects: Vec<String>, ) -> Result<(), ExogenesisError>

add new objects data

fn read_years(&self) -> Result<HashSet<String>, ExogenesisError>

fn read_months(&self, year: &str) -> Result<HashSet<String>, ExogenesisError>

fn read_days( &self, year: &str, month: &str, ) -> Result<HashSet<String>, ExogenesisError>

fn read_objects( &self, year: &str, month: &str, day: &str, ) -> Result<HashSet<String>, ExogenesisError>