Function app::touch_sensor::start_touch_sensor

pub fn start_touch_sensor() -> MynewtResult<()>

Initialise the touch controller. NFC antenna pins must already be reassigned as GPIO pins: Set NFC_PINS_AS_GPIO: 1 in hw/bsp/nrf52/syscfg.yml. To check whether whether NFC antenna pins have been correctly reassigned as GPIO pins, use the nrf52 crate and check that the output is fe:

let peripherals = nrf52::Peripherals::take().unwrap();
let nfcpins = peripherals.UICR.nfcpins.read().bits();
console::print("nfcpins = "); console::printhex(nfcpins as u8); console::print("\n");