Module app::uart

UART HAL

Structs

UART_CFG_Type

@brief UART configuration structure type definition

UART_FifoCfg_Type

@brief UART FIFO configuration structure type definition

UART_IrCfg_Type

@brief UART infrared configuration structure type definition

__BindgenBitfieldUnit

uart_reg

uart_reg__bindgen_ty_1__bindgen_ty_1

uart_reg__bindgen_ty_2__bindgen_ty_1

uart_reg__bindgen_ty_3__bindgen_ty_1

uart_reg__bindgen_ty_4__bindgen_ty_1

uart_reg__bindgen_ty_5__bindgen_ty_1

uart_reg__bindgen_ty_6__bindgen_ty_1

uart_reg__bindgen_ty_7__bindgen_ty_1

uart_reg__bindgen_ty_8__bindgen_ty_1

uart_reg__bindgen_ty_9__bindgen_ty_1

uart_reg__bindgen_ty_10__bindgen_ty_1

uart_reg__bindgen_ty_11__bindgen_ty_1

uart_reg__bindgen_ty_12__bindgen_ty_1

uart_reg__bindgen_ty_13__bindgen_ty_1

uart_reg__bindgen_ty_14__bindgen_ty_1

uart_reg__bindgen_ty_15__bindgen_ty_1

uart_reg__bindgen_ty_16__bindgen_ty_1

uart_reg__bindgen_ty_17__bindgen_ty_1

Constants

BL_Err_Type_ERROR

BL_Err_Type_SUCCESS

BL_Err_Type_TIMEOUT

BL_Fun_Type_DISABLE

BL_Fun_Type_ENABLE

BL_Mask_Type_MASK

BL_Mask_Type_UNMASK

BL_Sts_Type_RESET

BL_Sts_Type_SET

BL_UART_BUFFER_SIZE_MASK

BL_UART_BUFFER_SIZE_MIN

UART_AutoBaudDetection_Type_UART_AUTOBAUD_0X55

< UART auto baudrate detection using codeword 0x55

UART_AutoBaudDetection_Type_UART_AUTOBAUD_STARTBIT

< UART auto baudrate detection using start bit

UART_BIT_PRD_OFFSET

UART_ByteBitInverse_Type_UART_LSB_FIRST

< UART each byte is send out LSB-first

UART_ByteBitInverse_Type_UART_MSB_FIRST

< UART each byte is send out MSB-first

UART_CR_UART_BIT_INV_LEN

UART_CR_UART_BIT_INV_MSK

UART_CR_UART_BIT_INV_POS

UART_CR_UART_BIT_INV_UMSK

UART_CR_URX_ABR_EN_LEN

UART_CR_URX_ABR_EN_MSK

UART_CR_URX_ABR_EN_POS

UART_CR_URX_ABR_EN_UMSK

UART_CR_URX_BIT_CNT_D_LEN

UART_CR_URX_BIT_CNT_D_MSK

UART_CR_URX_BIT_CNT_D_POS

UART_CR_URX_BIT_CNT_D_UMSK

UART_CR_URX_BIT_PRD_LEN

UART_CR_URX_BIT_PRD_MSK

UART_CR_URX_BIT_PRD_POS

UART_CR_URX_BIT_PRD_UMSK

UART_CR_URX_DEG_CNT_LEN

UART_CR_URX_DEG_CNT_MSK

UART_CR_URX_DEG_CNT_POS

UART_CR_URX_DEG_CNT_UMSK

UART_CR_URX_DEG_EN_LEN

UART_CR_URX_DEG_EN_MSK

UART_CR_URX_DEG_EN_POS

UART_CR_URX_DEG_EN_UMSK

UART_CR_URX_END_CLR_LEN

UART_CR_URX_END_CLR_MSK

UART_CR_URX_END_CLR_POS

UART_CR_URX_END_CLR_UMSK

UART_CR_URX_END_EN_LEN

UART_CR_URX_END_EN_MSK

UART_CR_URX_END_EN_POS

UART_CR_URX_END_EN_UMSK

UART_CR_URX_END_MASK_LEN

UART_CR_URX_END_MASK_MSK

UART_CR_URX_END_MASK_POS

UART_CR_URX_END_MASK_UMSK

UART_CR_URX_EN_LEN

UART_CR_URX_EN_MSK

UART_CR_URX_EN_POS

UART_CR_URX_EN_UMSK

UART_CR_URX_FER_EN_LEN

UART_CR_URX_FER_EN_MSK

UART_CR_URX_FER_EN_POS

UART_CR_URX_FER_EN_UMSK

UART_CR_URX_FER_MASK_LEN

UART_CR_URX_FER_MASK_MSK

UART_CR_URX_FER_MASK_POS

UART_CR_URX_FER_MASK_UMSK

UART_CR_URX_FIFO_EN_LEN

UART_CR_URX_FIFO_EN_MSK

UART_CR_URX_FIFO_EN_POS

UART_CR_URX_FIFO_EN_UMSK

UART_CR_URX_FIFO_MASK_LEN

UART_CR_URX_FIFO_MASK_MSK

UART_CR_URX_FIFO_MASK_POS

UART_CR_URX_FIFO_MASK_UMSK

UART_CR_URX_IR_EN_LEN

UART_CR_URX_IR_EN_MSK

UART_CR_URX_IR_EN_POS

UART_CR_URX_IR_EN_UMSK

UART_CR_URX_IR_INV_LEN

UART_CR_URX_IR_INV_MSK

UART_CR_URX_IR_INV_POS

UART_CR_URX_IR_INV_UMSK

UART_CR_URX_IR_POS_S_LEN

UART_CR_URX_IR_POS_S_MSK

UART_CR_URX_IR_POS_S_POS

UART_CR_URX_IR_POS_S_UMSK

UART_CR_URX_LEN_LEN

UART_CR_URX_LEN_MSK

UART_CR_URX_LEN_POS

UART_CR_URX_LEN_UMSK

UART_CR_URX_PCE_CLR_LEN

UART_CR_URX_PCE_CLR_MSK

UART_CR_URX_PCE_CLR_POS

UART_CR_URX_PCE_CLR_UMSK

UART_CR_URX_PCE_EN_LEN

UART_CR_URX_PCE_EN_MSK

UART_CR_URX_PCE_EN_POS

UART_CR_URX_PCE_EN_UMSK

UART_CR_URX_PCE_MASK_LEN

UART_CR_URX_PCE_MASK_MSK

UART_CR_URX_PCE_MASK_POS

UART_CR_URX_PCE_MASK_UMSK

UART_CR_URX_PRT_EN_LEN

UART_CR_URX_PRT_EN_MSK

UART_CR_URX_PRT_EN_POS

UART_CR_URX_PRT_EN_UMSK

UART_CR_URX_PRT_SEL_LEN

UART_CR_URX_PRT_SEL_MSK

UART_CR_URX_PRT_SEL_POS

UART_CR_URX_PRT_SEL_UMSK

UART_CR_URX_RTO_CLR_LEN

UART_CR_URX_RTO_CLR_MSK

UART_CR_URX_RTO_CLR_POS

UART_CR_URX_RTO_CLR_UMSK

UART_CR_URX_RTO_EN_LEN

UART_CR_URX_RTO_EN_MSK

UART_CR_URX_RTO_EN_POS

UART_CR_URX_RTO_EN_UMSK

UART_CR_URX_RTO_MASK_LEN

UART_CR_URX_RTO_MASK_MSK

UART_CR_URX_RTO_MASK_POS

UART_CR_URX_RTO_MASK_UMSK

UART_CR_URX_RTO_VALUE_LEN

UART_CR_URX_RTO_VALUE_MSK

UART_CR_URX_RTO_VALUE_POS

UART_CR_URX_RTO_VALUE_UMSK

UART_CR_URX_RTS_SW_MODE_LEN

UART_CR_URX_RTS_SW_MODE_MSK

UART_CR_URX_RTS_SW_MODE_POS

UART_CR_URX_RTS_SW_MODE_UMSK

UART_CR_URX_RTS_SW_VAL_LEN

UART_CR_URX_RTS_SW_VAL_MSK

UART_CR_URX_RTS_SW_VAL_POS

UART_CR_URX_RTS_SW_VAL_UMSK

UART_CR_UTX_BIT_CNT_D_LEN

UART_CR_UTX_BIT_CNT_D_MSK

UART_CR_UTX_BIT_CNT_D_POS

UART_CR_UTX_BIT_CNT_D_UMSK

UART_CR_UTX_BIT_CNT_P_LEN

UART_CR_UTX_BIT_CNT_P_MSK

UART_CR_UTX_BIT_CNT_P_POS

UART_CR_UTX_BIT_CNT_P_UMSK

UART_CR_UTX_BIT_PRD_LEN

UART_CR_UTX_BIT_PRD_MSK

UART_CR_UTX_BIT_PRD_POS

UART_CR_UTX_BIT_PRD_UMSK

UART_CR_UTX_CTS_EN_LEN

UART_CR_UTX_CTS_EN_MSK

UART_CR_UTX_CTS_EN_POS

UART_CR_UTX_CTS_EN_UMSK

UART_CR_UTX_END_CLR_LEN

UART_CR_UTX_END_CLR_MSK

UART_CR_UTX_END_CLR_POS

UART_CR_UTX_END_CLR_UMSK

UART_CR_UTX_END_EN_LEN

UART_CR_UTX_END_EN_MSK

UART_CR_UTX_END_EN_POS

UART_CR_UTX_END_EN_UMSK

UART_CR_UTX_END_MASK_LEN

UART_CR_UTX_END_MASK_MSK

UART_CR_UTX_END_MASK_POS

UART_CR_UTX_END_MASK_UMSK

UART_CR_UTX_EN_LEN

UART_CR_UTX_EN_MSK

UART_CR_UTX_EN_POS

UART_CR_UTX_EN_UMSK

UART_CR_UTX_FER_EN_LEN

UART_CR_UTX_FER_EN_MSK

UART_CR_UTX_FER_EN_POS

UART_CR_UTX_FER_EN_UMSK

UART_CR_UTX_FER_MASK_LEN

UART_CR_UTX_FER_MASK_MSK

UART_CR_UTX_FER_MASK_POS

UART_CR_UTX_FER_MASK_UMSK

UART_CR_UTX_FIFO_EN_LEN

UART_CR_UTX_FIFO_EN_MSK

UART_CR_UTX_FIFO_EN_POS

UART_CR_UTX_FIFO_EN_UMSK

UART_CR_UTX_FIFO_MASK_LEN

UART_CR_UTX_FIFO_MASK_MSK

UART_CR_UTX_FIFO_MASK_POS

UART_CR_UTX_FIFO_MASK_UMSK

UART_CR_UTX_FRM_EN_LEN

UART_CR_UTX_FRM_EN_MSK

UART_CR_UTX_FRM_EN_POS

UART_CR_UTX_FRM_EN_UMSK

UART_CR_UTX_IR_EN_LEN

UART_CR_UTX_IR_EN_MSK

UART_CR_UTX_IR_EN_POS

UART_CR_UTX_IR_EN_UMSK

UART_CR_UTX_IR_INV_LEN

UART_CR_UTX_IR_INV_MSK

UART_CR_UTX_IR_INV_POS

UART_CR_UTX_IR_INV_UMSK

UART_CR_UTX_IR_POS_P_LEN

UART_CR_UTX_IR_POS_P_MSK

UART_CR_UTX_IR_POS_P_POS

UART_CR_UTX_IR_POS_P_UMSK

UART_CR_UTX_IR_POS_S_LEN

UART_CR_UTX_IR_POS_S_MSK

UART_CR_UTX_IR_POS_S_POS

UART_CR_UTX_IR_POS_S_UMSK

UART_CR_UTX_LEN_LEN

UART_CR_UTX_LEN_MSK

UART_CR_UTX_LEN_POS

UART_CR_UTX_LEN_UMSK

UART_CR_UTX_PRT_EN_LEN

UART_CR_UTX_PRT_EN_MSK

UART_CR_UTX_PRT_EN_POS

UART_CR_UTX_PRT_EN_UMSK

UART_CR_UTX_PRT_SEL_LEN

UART_CR_UTX_PRT_SEL_MSK

UART_CR_UTX_PRT_SEL_POS

UART_CR_UTX_PRT_SEL_UMSK

UART_DATA_CONFIG_OFFSET

UART_DEFAULT_RECV_TIMEOUT

UART_DMA_RX_EN_LEN

UART_DMA_RX_EN_MSK

UART_DMA_RX_EN_POS

UART_DMA_RX_EN_UMSK

UART_DMA_TX_EN_LEN

UART_DMA_TX_EN_MSK

UART_DMA_TX_EN_POS

UART_DMA_TX_EN_UMSK

UART_DataBits_Type_UART_DATABITS_5

< UART data bits length:5 bits

UART_DataBits_Type_UART_DATABITS_6

< UART data bits length:6 bits

UART_DataBits_Type_UART_DATABITS_7

< UART data bits length:7 bits

UART_DataBits_Type_UART_DATABITS_8

< UART data bits length:8 bits

UART_Direction_Type_UART_RX

< UART RX Direction

UART_Direction_Type_UART_TX

< UART TX Direction

UART_Direction_Type_UART_TXRX

< UART TX and RX Direction

UART_FIFO_CONFIG_0_OFFSET

UART_FIFO_CONFIG_1_OFFSET

UART_FIFO_RDATA_LEN

UART_FIFO_RDATA_MSK

UART_FIFO_RDATA_OFFSET

UART_FIFO_RDATA_POS

UART_FIFO_RDATA_UMSK

UART_FIFO_WDATA_LEN

UART_FIFO_WDATA_MSK

UART_FIFO_WDATA_OFFSET

UART_FIFO_WDATA_POS

UART_FIFO_WDATA_UMSK

UART_ID_Type_UART0_ID

< UART0 port define

UART_ID_Type_UART1_ID

< UART1 port define

UART_ID_Type_UART_ID_MAX

< UART MAX ID define

UART_INT_CLEAR_OFFSET

UART_INT_EN_OFFSET

UART_INT_MASK_OFFSET

UART_INT_STS_OFFSET

UART_INT_Type_UART_INT_ALL

< All the interrupt

UART_INT_Type_UART_INT_PCE

< UART rx parity check error interrupt

UART_INT_Type_UART_INT_RTO

< UART rx time-out interrupt

UART_INT_Type_UART_INT_RX_END

< UART rx transfer end interrupt

UART_INT_Type_UART_INT_RX_FER

< UART rx fifo overflow/underflow error interrupt

UART_INT_Type_UART_INT_RX_FIFO_REQ

< UART rx fifo interrupt when rx fifo count reaches,auto clear

UART_INT_Type_UART_INT_TX_END

< UART tx transfer end interrupt

UART_INT_Type_UART_INT_TX_FER

< UART tx fifo overflow/underflow error interrupt

UART_INT_Type_UART_INT_TX_FIFO_REQ

< UART tx fifo interrupt when tx fifo count reaches,auto clear

UART_Overflow_Type_UART_RX_OVERFLOW

< UART rx fifo overflow

UART_Overflow_Type_UART_RX_UNDERFLOW

< UART rx fifo underflow

UART_Overflow_Type_UART_TX_OVERFLOW

< UART tx fifo overflow

UART_Overflow_Type_UART_TX_UNDERFLOW

< UART tx fifo underflow

UART_Parity_Type_UART_PARITY_EVEN

< UART parity even define

UART_Parity_Type_UART_PARITY_NONE

< UART parity none define

UART_Parity_Type_UART_PARITY_ODD

< UART parity odd define

UART_RX_FIFO_CLR_LEN

UART_RX_FIFO_CLR_MSK

UART_RX_FIFO_CLR_POS

UART_RX_FIFO_CLR_UMSK

UART_RX_FIFO_CNT_LEN

UART_RX_FIFO_CNT_MSK

UART_RX_FIFO_CNT_POS

UART_RX_FIFO_CNT_UMSK

UART_RX_FIFO_OVERFLOW_LEN

UART_RX_FIFO_OVERFLOW_MSK

UART_RX_FIFO_OVERFLOW_POS

UART_RX_FIFO_OVERFLOW_UMSK

UART_RX_FIFO_SIZE

UART_RX_FIFO_TH_LEN

UART_RX_FIFO_TH_MSK

UART_RX_FIFO_TH_POS

UART_RX_FIFO_TH_UMSK

UART_RX_FIFO_UNDERFLOW_LEN

UART_RX_FIFO_UNDERFLOW_MSK

UART_RX_FIFO_UNDERFLOW_POS

UART_RX_FIFO_UNDERFLOW_UMSK

UART_STATUS_OFFSET

UART_STS_URX_ABR_PRD_0X55_LEN

UART_STS_URX_ABR_PRD_0X55_MSK

UART_STS_URX_ABR_PRD_0X55_POS

UART_STS_URX_ABR_PRD_0X55_UMSK

UART_STS_URX_ABR_PRD_OFFSET

UART_STS_URX_ABR_PRD_START_LEN

UART_STS_URX_ABR_PRD_START_MSK

UART_STS_URX_ABR_PRD_START_POS

UART_STS_URX_ABR_PRD_START_UMSK

UART_STS_URX_BUS_BUSY_LEN

UART_STS_URX_BUS_BUSY_MSK

UART_STS_URX_BUS_BUSY_POS

UART_STS_URX_BUS_BUSY_UMSK

UART_STS_UTX_BUS_BUSY_LEN

UART_STS_UTX_BUS_BUSY_MSK

UART_STS_UTX_BUS_BUSY_POS

UART_STS_UTX_BUS_BUSY_UMSK

UART_StopBits_Type_UART_STOPBITS_1

< UART data stop bits length:1 bits

UART_StopBits_Type_UART_STOPBITS_1_5

< UART data stop bits length:1.5 bits

UART_StopBits_Type_UART_STOPBITS_2

< UART data stop bits length:2 bits

UART_TX_FIFO_CLR_LEN

UART_TX_FIFO_CLR_MSK

UART_TX_FIFO_CLR_POS

UART_TX_FIFO_CLR_UMSK

UART_TX_FIFO_CNT_LEN

UART_TX_FIFO_CNT_MSK

UART_TX_FIFO_CNT_POS

UART_TX_FIFO_CNT_UMSK

UART_TX_FIFO_OVERFLOW_LEN

UART_TX_FIFO_OVERFLOW_MSK

UART_TX_FIFO_OVERFLOW_POS

UART_TX_FIFO_OVERFLOW_UMSK

UART_TX_FIFO_SIZE

UART_TX_FIFO_TH_LEN

UART_TX_FIFO_TH_MSK

UART_TX_FIFO_TH_POS

UART_TX_FIFO_TH_UMSK

UART_TX_FIFO_UNDERFLOW_LEN

UART_TX_FIFO_UNDERFLOW_MSK

UART_TX_FIFO_UNDERFLOW_POS

UART_TX_FIFO_UNDERFLOW_UMSK

UART_URX_CONFIG_OFFSET

UART_URX_END_INT_LEN

UART_URX_END_INT_MSK

UART_URX_END_INT_POS

UART_URX_END_INT_UMSK

UART_URX_FER_INT_LEN

UART_URX_FER_INT_MSK

UART_URX_FER_INT_POS

UART_URX_FER_INT_UMSK

UART_URX_FIFO_INT_LEN

UART_URX_FIFO_INT_MSK

UART_URX_FIFO_INT_POS

UART_URX_FIFO_INT_UMSK

UART_URX_IR_POSITION_OFFSET

UART_URX_PCE_INT_LEN

UART_URX_PCE_INT_MSK

UART_URX_PCE_INT_POS

UART_URX_PCE_INT_UMSK

UART_URX_RTO_INT_LEN

UART_URX_RTO_INT_MSK

UART_URX_RTO_INT_POS

UART_URX_RTO_INT_UMSK

UART_URX_RTO_TIMER_OFFSET

UART_UTX_CONFIG_OFFSET

UART_UTX_END_INT_LEN

UART_UTX_END_INT_MSK

UART_UTX_END_INT_POS

UART_UTX_END_INT_UMSK

UART_UTX_FER_INT_LEN

UART_UTX_FER_INT_MSK

UART_UTX_FER_INT_POS

UART_UTX_FER_INT_UMSK

UART_UTX_FIFO_INT_LEN

UART_UTX_FIFO_INT_MSK

UART_UTX_FIFO_INT_POS

UART_UTX_FIFO_INT_UMSK

UART_UTX_IR_POSITION_OFFSET

Functions

UART_AutoBaudDetection⚠

UART_ClrRtsValue⚠

UART_DeInit⚠

UART_Disable⚠

UART_Enable⚠

UART_FifoConfig⚠

UART_GetAutoBaudCount⚠

UART_GetIntStatus⚠

UART_GetOverflowStatus⚠

UART_GetRxBusBusyStatus⚠

UART_GetRxFifoCount⚠

UART_GetTxBusBusyStatus⚠

UART_GetTxFifoCount⚠

UART_Init⚠

UART_IntClear⚠

UART_IntMask⚠

UART_Int_Callback_Install⚠

UART_IrConfig⚠

UART_ReceiveData⚠

UART_RxFifoClear⚠

UART_SendData⚠

UART_SendDataBlock⚠

UART_SetBaudrate⚠

UART_SetDeglitchCount⚠

UART_SetRtsValue⚠

UART_SetRxDataLength⚠

UART_SetRxTimeoutValue⚠

UART_SetTxDataLength⚠

UART_TxFifoClear⚠

UART_TxFreeRun⚠

uart_data_recv

uart_data_send

uart_datas_send

uart_debug_early_init

uart_early_init

uart_flush

uart_getdefconfig

uart_gpio_init

uart_init

uart_int_disable

uart_int_enable

uart_int_rx_disable

uart_int_rx_enable

uart_int_rx_notify_register

uart_int_rx_notify_unregister

uart_int_tx_disable

uart_int_tx_enable

uart_int_tx_notify_register

uart_int_tx_notify_unregister

uart_setbaud

uart_setconfig

uart_string_send

Type Definitions

BL_Err_Type

@brief Error type definition

BL_Fun_Type

@brief Functional type definition

BL_Mask_Type

@brief Mask type definition

BL_Sts_Type

@brief Status type definition

UART_AutoBaudDetection_Type

@brief UART auto baudrate detection using codeword 0x55 or start bit definiton

UART_ByteBitInverse_Type

@brief UART each data byte is send out LSB-first or MSB-first type definiton

UART_DataBits_Type

@brief UART data bits type definiton

UART_Direction_Type

@brief UART direction type definition

UART_ID_Type

@brief UART port type definition

UART_INT_Type

@brief UART interrupt type definition

UART_Overflow_Type

@brief UART overflow or underflow type definition

UART_Parity_Type

@brief UART parity type definition

UART_StopBits_Type

@brief UART stop bits type definiton

__uint8_t

__uint16_t

__uint32_t

cb_uart_notify_t

intCallback_Type

@brief Interrupt callback function type

uart_reg_t

Unions

uart_reg__bindgen_ty_1

uart_reg__bindgen_ty_2

uart_reg__bindgen_ty_3

uart_reg__bindgen_ty_4

uart_reg__bindgen_ty_5

uart_reg__bindgen_ty_6

uart_reg__bindgen_ty_7

uart_reg__bindgen_ty_8

uart_reg__bindgen_ty_9

uart_reg__bindgen_ty_10

uart_reg__bindgen_ty_11

uart_reg__bindgen_ty_12

uart_reg__bindgen_ty_13

uart_reg__bindgen_ty_14

uart_reg__bindgen_ty_15

uart_reg__bindgen_ty_16

uart_reg__bindgen_ty_17