Module app::wifi

Re-exports

pub use self::WIFI_MGMR_EVENT as WIFI_MGMR_EVENT_T;

pub use self::WIFI_MGMR_CONNECTION_STATUS as WIFI_MGMR_CONNECTION_STATUS_T;

Structs

__BindgenBitfieldUnit

__IncompleteArrayField

event

\brief Event

ip4_addr

This is the aligned version of ip4_addr_t, used as local variable, on the stack, etc.

netif

Generic data structure used for all lwIP network interfaces. The following fields should be filled in by the initialization function for the device driver: hwaddr_len, hwaddr[], mtu, flags

pbuf

Main packet buffer struct

state

\brief State

stateMachine

\brief State machine

transition

\brief Transition between a state and another state

wifi_conf

wifi_mgmr

wifi_mgmr_ap_item

wifi_mgmr_ap_msg

wifi_mgmr_cfg_element_msg

wifi_mgmr_cipher_t

wifi_mgmr_connect_ind_stat_info

wifi_mgmr_ipgot_msg

wifi_mgmr_msg

wifi_mgmr_profile

wifi_mgmr_profile_msg

wifi_mgmr_scan_item

wifi_mgmr_sta_basic_info

wifi_mgmr_sta_connect_ind_stat_info

wifi_sta_basic_info

wlan_netif

wlan_netif__bindgen_ty_1

wlan_netif__bindgen_ty_2__bindgen_ty_1

xSTATIC_LIST_ITEM

xSTATIC_STREAM_BUFFER

xSTATIC_TIMER

Constants

WIFI_MGMR_CONFIG_SCAN_ITEM_TIMEOUT

WIFI_MGMR_CONNECTION_STATUS_WIFI_MGMR_CONNECTION_STATUS_CONNECTED_IP_NO

WIFI_MGMR_CONNECTION_STATUS_WIFI_MGMR_CONNECTION_STATUS_CONNECTED_IP_YES

WIFI_MGMR_CONNECTION_STATUS_WIFI_MGMR_CONNECTION_STATUS_CONNECTING

WIFI_MGMR_CONNECTION_STATUS_WIFI_MGMR_CONNECTION_STATUS_DISCONNECTED

WIFI_MGMR_CONNECTION_STATUS_WIFI_MGMR_CONNECTION_STATUS_IDLE

WIFI_MGMR_CONNECT_IND_STAT_INFO_TYPE_IND_CONNECTION

WIFI_MGMR_CONNECT_IND_STAT_INFO_TYPE_IND_DISCONNECTION

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_AP_START

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_AP_STOP

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_CONF_MAX_STA

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_CONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_CONNECTED

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_DENOISE

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_DISCONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_IDLE

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_IP_GOT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_PHY_UP

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_RC_CONFIG

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_RECONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_RELOAD_TSEN

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_APP_SNIFFER

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_CFG_REQ

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_CHANNEL_SET

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_DATA_RAW_SEND

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_DISCONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_IND_CONNECTED

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_IND_DISCONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_POWERSAVING

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_FW_SCAN

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_AP_IND_STA_DEL

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_AP_IND_STA_NEW

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_DISABLE_AUTORECONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_ENABLE_AUTORECONNECT

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_IP_UPDATE

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_SCAN_IND_BEACON

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_GLB_SCAN_IND_PROBE_RESP

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_MAXAPP_MINIFW

WIFI_MGMR_EVENT_WIFI_MGMR_EVENT_MAXFW_MINI_GLOBAL

WIFI_MGMR_FEATURES_SCAN_SAVE_HIDDEN_SSID

WIFI_MGMR_MQ_MSG_COUNT

WIFI_MGMR_MQ_MSG_SIZE

WIFI_MGMR_PENDING_TASK_SCAN_BIT

WIFI_MGMR_PROFILES_MAX

WIFI_MGMR_SCAN_ITEMS_MAX

WIFI_SCAN_DONE_EVENT_TYPE_WIFI_SCAN_DONE_EVENT_BUSY

WIFI_SCAN_DONE_EVENT_TYPE_WIFI_SCAN_DONE_EVENT_OK

WIFI_STATE_ENUM_LIST_WIFI_STATE_CONNECTED_IP_GETTING

WIFI_STATE_ENUM_LIST_WIFI_STATE_CONNECTED_IP_GOT

WIFI_STATE_ENUM_LIST_WIFI_STATE_CONNECTING

WIFI_STATE_ENUM_LIST_WIFI_STATE_DISCONNECT

WIFI_STATE_ENUM_LIST_WIFI_STATE_IDLE

WIFI_STATE_ENUM_LIST_WIFI_STATE_IFDOWN

WIFI_STATE_ENUM_LIST_WIFI_STATE_NO_AP_FOUND

WIFI_STATE_ENUM_LIST_WIFI_STATE_PSK_ERROR

WIFI_STATE_ENUM_LIST_WIFI_STATE_SNIFFER

WIFI_STATE_ENUM_LIST_WIFI_STATE_UNKNOWN

WIFI_STATE_ENUM_LIST_WIFI_STATE_WITH_AP_CONNECTED_IP_GETTING

WIFI_STATE_ENUM_LIST_WIFI_STATE_WITH_AP_CONNECTED_IP_GOT

WIFI_STATE_ENUM_LIST_WIFI_STATE_WITH_AP_CONNECTING

WIFI_STATE_ENUM_LIST_WIFI_STATE_WITH_AP_DISCONNECT

WIFI_STATE_ENUM_LIST_WIFI_STATE_WITH_AP_IDLE

netif_mac_filter_action_NETIF_ADD_MAC_FILTER

Add a filter entry

netif_mac_filter_action_NETIF_DEL_MAC_FILTER

Delete a filter entry

Statics

wifiMgmr

Functions

wifi_mgmr_all_ap_scan⚠

wifi_mgmr_ap_enable⚠

wifi_mgmr_ap_ip_get⚠

wifi_mgmr_ap_mac_get⚠

wifi_mgmr_ap_mac_set⚠

wifi_mgmr_ap_set_gateway⚠

wifi_mgmr_ap_sta_cnt_get⚠

wifi_mgmr_ap_sta_cnt_get_internal⚠

wifi_mgmr_ap_sta_delete⚠

wifi_mgmr_ap_sta_delete_internal⚠

wifi_mgmr_ap_sta_info_get⚠

wifi_mgmr_ap_sta_info_get_internal⚠

wifi_mgmr_ap_start⚠

wifi_mgmr_ap_stop⚠

wifi_mgmr_api_fw_tsen_reload⚠

wifi_mgmr_auth_to_str⚠

wifi_mgmr_cfg_req⚠

wifi_mgmr_channel_get⚠

wifi_mgmr_channel_set⚠

wifi_mgmr_cipher_to_str⚠

wifi_mgmr_cli_init⚠

wifi_mgmr_cli_scanlist⚠

wifi_mgmr_conf_max_sta⚠

wifi_mgmr_drv_init⚠

wifi_mgmr_event_notify⚠

wifi_mgmr_ext_dump_needed⚠

wifi_mgmr_get_wifi_channel_conf⚠

wifi_mgmr_init⚠

wifi_mgmr_psk_cal⚠

wifi_mgmr_rate_config⚠

wifi_mgmr_raw_80211_send⚠

wifi_mgmr_rssi_get⚠

wifi_mgmr_scan⚠

wifi_mgmr_scan_ap⚠

wifi_mgmr_scan_ap_all⚠

wifi_mgmr_scan_complete_callback⚠

wifi_mgmr_scan_complete_notify⚠

wifi_mgmr_scan_filter_hidden_ssid⚠

wifi_mgmr_scan_item_is_timeout⚠

wifi_mgmr_set_country_code⚠

wifi_mgmr_set_country_code_internal⚠

wifi_mgmr_sniffer_disable⚠

wifi_mgmr_sniffer_enable⚠

wifi_mgmr_sniffer_register⚠

wifi_mgmr_sniffer_unregister⚠

wifi_mgmr_sta_autoconnect_disable⚠

wifi_mgmr_sta_autoconnect_enable⚠

wifi_mgmr_sta_connect⚠

wifi_mgmr_sta_connect_ind_stat_get⚠

wifi_mgmr_sta_disable⚠

wifi_mgmr_sta_disconnect⚠

wifi_mgmr_sta_dns_get⚠

wifi_mgmr_sta_enable⚠

wifi_mgmr_sta_ip_get⚠

wifi_mgmr_sta_ip_set⚠

wifi_mgmr_sta_ip_unset⚠

wifi_mgmr_sta_mac_get⚠

wifi_mgmr_sta_mac_set⚠

wifi_mgmr_sta_powersaving⚠

wifi_mgmr_sta_psk_set⚠

wifi_mgmr_sta_ssid_set⚠

wifi_mgmr_start⚠

wifi_mgmr_start_background⚠

wifi_mgmr_state_get⚠

wifi_mgmr_state_get_internal⚠

wifi_mgmr_status_code_clean_internal⚠

wifi_mgmr_status_code_get⚠

wifi_mgmr_status_code_get_internal⚠

wifi_mgmr_status_code_str⚠

wifi_start_firmware_task

Type Definitions

StaticListItem_t

StaticMessageBuffer_t

StaticStreamBuffer_t

StaticTimer_t

TaskFunction_t

TickType_t

UBaseType_t

WIFI_MGMR_CONNECTION_STATUS

WIFI_MGMR_EVENT

WIFI_SCAN_DONE_EVENT_TYPE

WIFI_STATE_ENUM_LIST

__int8_t

__int32_t

__uint8_t

__uint16_t

__uint32_t

err_t

ip4_addr_t

ip4_addr_t uses a struct for convenience only, so that the same defines can operate both on ip4_addr_t as well as on ip4_addr_p_t.

ip_addr_t

netif_igmp_mac_filter_fn

Function prototype for netif igmp_mac_filter functions

netif_input_fn

Function prototype for netif->input functions. This function is saved as ‘input’ callback function in the netif struct. Call it when a packet has been received.

netif_linkoutput_fn

Function prototype for netif->linkoutput functions. Only used for ethernet netifs. This function is called by ARP when a packet shall be sent.

netif_mac_filter_action

MAC Filter Actions, these are passed to a netif’s igmp_mac_filter or mld_mac_filter callback function.

netif_output_fn

Function prototype for netif->output functions. Called by lwIP when a packet shall be sent. For ethernet netif, set this to ‘etharp_output’ and set ‘linkoutput’.

netif_status_callback_fn

Function prototype for netif status- or link-callback functions.

os_messagequeue_t

os_timer_t

s8_t

scan_complete_cb_t

scan_item_cb_t

size_t

sniffer_cb_t

u8_t

u16_t

u32_t

wifi_conf_t

wifi_interface_t

wifi_mgmr_ap_item_t

wifi_mgmr_ap_msg_t

wifi_mgmr_cfg_element_msg_t

wifi_mgmr_connect_ind_stat_info_t

wifi_mgmr_ipgot_msg_t

wifi_mgmr_msg_t

wifi_mgmr_profile_msg_t

wifi_mgmr_profile_t

wifi_mgmr_scan_item_t

wifi_mgmr_sta_basic_info_t

wifi_mgmr_sta_connect_ind_stat_info_t

wifi_mgmr_t

wifi_sta_basic_info_t

Unions

wlan_netif__bindgen_ty_2