Struct app::wifi::netif

#[repr(C)]
pub struct netif {

    pub next: *mut netif,
    pub ip_addr: ip_addr_t,
    pub netmask: ip_addr_t,
    pub gw: ip_addr_t,
    pub input: netif_input_fn,
    pub output: netif_output_fn,
    pub linkoutput: netif_linkoutput_fn,
    pub status_callback: netif_status_callback_fn,
    pub link_callback: netif_status_callback_fn,
    pub state: *mut c_void,
    pub client_data: [*mut c_void; 3],
    pub hostname: *const c_char,
    pub mtu: u16_t,
    pub hwaddr: [u8_t; 6],
    pub hwaddr_len: u8_t,
    pub flags: u8_t,
    pub name: [c_char; 2],
    pub num: u8_t,
    pub rs_count: u8_t,
    pub igmp_mac_filter: netif_igmp_mac_filter_fn,

}

Generic data structure used for all lwIP network interfaces. The following fields should be filled in by the initialization function for the device driver: hwaddr_len, hwaddr[], mtu, flags

Fields
next: *mut netif

pointer to next in linked list

ip_addr: ip_addr_t

IP address configuration in network byte order

netmask: ip_addr_tgw: ip_addr_tinput: netif_input_fn

This function is called by the network device driver to pass a packet up the TCP/IP stack.

output: netif_output_fn

This function is called by the IP module when it wants to send a packet on the interface. This function typically first resolves the hardware address, then sends the packet. For ethernet physical layer, this is usually etharp_output()

linkoutput: netif_linkoutput_fn

This function is called by ethernet_output() when it wants to send a packet on the interface. This function outputs the pbuf as-is on the link medium.

status_callback: netif_status_callback_fn

This function is called when the netif state is set to up or down

link_callback: netif_status_callback_fn

This function is called when the netif link is set to up or down

state: *mut c_void

This field can be set by the device driver and could point to state information for the device.

client_data: [*mut c_void; 3]hostname: *const c_charmtu: u16_t

maximum transfer unit (in bytes)

hwaddr: [u8_t; 6]

link level hardware address of this interface

hwaddr_len: u8_t

number of bytes used in hwaddr

flags: u8_t

flags (@see @ref netif_flags)

name: [c_char; 2]

descriptive abbreviation

num: u8_t

number of this interface. Used for @ref if_api and @ref netifapi_netif, as well as for IPv6 zones

rs_count: u8_t

Number of Router Solicitation messages that remain to be sent.

igmp_mac_filter: netif_igmp_mac_filter_fn

This function could be called to add or delete an entry in the multicast filter table of the ethernet MAC.